[Fix] 네트워크 소켓 연동 수정

+++ Assets/WorkSpaces/WGH/Scripts/NPC/WGH_NPCCreator.cs
@@ -56,7 +56,7 @@ public class WGH_NPCCreator : MonoBehaviour
         storeNpcSpawnTime = SetCustomerAmount(stageLevel);
     }
 
-    private int SetCustomerAmount(int stageLevel)
+    private float SetCustomerAmount(int stageLevel)
     {
         switch(stageLevel)
         {
